Move RawDataProduct class into its own files.
